Creates a cypher table from a Numenera sourcebook
Convert the relevant sourcebook from .pdf to .txt and then adjust as-appropriate to strip extraneous content and convert the regularly-formatted cyphers or artifacts into a table that can be easily referenced or pulled from. Note that the output will need some editing due to the irregularities caused by the .pdf-to-.txt conversion process.